Allow more than 2 measurements per entity to be stored
TL;DR: gnocchi timeseries deserialization logic does not
       account for multi-valued aggregation methods

On POSTing the third measurement to an entity, the creation of
the pandas Timestamp from the deseralized blob retrieved from
swift fails with:

    ValueError: Unable to parse high

when attempting to parse what it expects to be a timestamp.

This occurs as a result of the open-high-low-close aggregate
not acting as expected in terms of producing a *single*
aggregated value per period. Instead there are 4 separate
aggregated valued involved: 'high', 'close', 'open' & 'low'.

So after the third datapoint is stored, the stored data for
ohlc is no longer just a simple dict mapping timestamps to
their corresponding values, as expected by the deserialization
logic.

Instead there are are also 'high', 'close', 'open' & 'low'
keys in the dict, e.g.:

  [{u'aggregation_method': u'ohlc',
    u'values': {u'high': 32.0,
                u'close': 32.0,
                u'open': 32.0,
                u'low': 32.0,
                u'2013-01-01 23:23:31+00:00': 31.0},
    u'max_size': 3600,
    u'sampling': u'60S'}]

For now, we remove 'ohlc' from the set of supported aggregation
methods, until we can figure out how to accomodate multi-valued
aggregation methods.

Change-Id: I81cac6e5d008346a78625e895271f3a61a3fb8ec