Remove aggregation method that requires full history retention
Accurately calculating aggregates such as quantiles requires that
the full history of measurements is retained.

This doesn't fit in well with the gnocchi approach of minimal
retention.

Currently the computed value for the median aggregate depends on
the order in which the measurements are submitted, i.e. permutation
of the datapoints yields a different result, which is clearly wrong.

For now, we remove 'median' from the set of supported aggregation
methods, until we can figure out how to accomodate history-dependent
aggregation methods.

Change-Id: I4355da2f1d6d06be9ff773cc77f1e076c245463c