Add support for selectable granularity
Instead of retrieving datapoints of mixed granularities from
multi-archive entities, the caller now has the option to specify
the required granularity in seconds.

For example to explicitly select per-minute means:

  GET /v1/entity/UUID/measures?aggregation=mean&granularity=60

If there is no archive of that granularity defined for the entity,
then the storage driver is free to return an empty response, or to
downsample to the requested granularity on-demand (in the case where
the driver has retained sufficient finer-grained data to do so).

Change-Id: I01d117e90bd352521a4e5e556cc09ed7fcefb367