Extension for moving aggregation in API layer
Added extension for moving average, moving variance,
ewma. Unit tests added.

Query syntax (recognizes two parameters, granularity and center):

    GET /v1/entity/UUID/measures?aggregation=moving-average&
	granularity=30S&center=False
or
    GET /v1/entity/UUID/measures?aggregation=ewma&granularity=60S

If granularity is not specified, the moving aggregates are computed
on the data using the coarsest archive granularity. If 1S resolution
data is not available, the moving aggregates are performed on the
rolled-up timeseries.

Change-Id: If40e58480f1512dd17b9c09cb75fd54c7193c257