Fix continuous test failure
The test_list_resources_by_unknown_field testcase is failing with the
new pecan version (0.6.0) with the NullIndexer. After creating a resource,
as a first step, it starts working, the test skipping logic is not failing
because of getting the standard HTTP Resource Not Found error from webob
instead of the NotImplemented error. The failure only occurs in case of this
test, the newly added negative test for get_resource works fine and also
the positive tests for list_resources.

Change-Id: If9d536a4060624a3baf37e082e397bb8afccb6fd