Fix continuous test failure and MySQL URL
The test_list_resources_by_unknown_field testcase is failing with the
new pecan version (0.6.0) with the NullIndexer. With PostgreSQL
all the list_resources related tests are failing. Testing against real
MySQL provides almost the same results, there are even more failing
testcases there. The URL for testing in run-mysql-tests.sh was fixed
to avoid skipping the tests on the gate.

After creating a resource, as a first step, in the
test_list_resources_by_unknown_field testcase it starts working, the
test skipping logic is not failing because of getting the standard
HTTP Resource Not Found error coming from webob instead of the
NotImplemented error.

The newly added negative test for get_resource works for PostgreSQL
acoording to the gate logs, but fails with MySQL.

Change-Id: If9d536a4060624a3baf37e082e397bb8afccb6fd