Fix continuous test failure with latest pecan
The test_list_resources_by_unknown_field testcase is failing with the
new pecan version (0.6.0) with the NullIndexer. With PostgreSQL
all the list_resources related tests are failing.

After creating a resource, as a first step, in the
test_list_resources_by_unknown_field testcase it starts working, the
test skipping logic is not failing because of getting the standard
HTTP Resource Not Found error coming from webob instead of the
NotImplemented error.

Change-Id: If9d536a4060624a3baf37e082e397bb8afccb6fd