Fix for unneccessary resampling of single-value timeseries
Resampling a timeseries with one element is unnecessary,
and causes problems in the case of an aggregated timeseries
with aggregation method set to stddev, as the single already-
aggregated value is then discarded due to standard deviation
being undefined for a single datapoint.

The new test fails with the original unselective resampling.

Change-Id: I7f0db8050c0cef5eceec24977e4cb02d4843824d