rest: enhance enforce context, fix created_by rule
In many places the context gave to enforce() is too minimal to apply
interesting rules, e.g. we need to pass the whole resource so we can
apply fine grained rules.

This also fixes created_by rule to match the created_by_project_id, not
the project_id as some entities such as metrics do not have these
properties anyway.

Change-Id: I49bd3c678831287a07f4389739e4b37ea845187b