tests: use non-admin user by default in test
By default we used to run query against the REST server in tests with an
admin user. This bypassed a lot of rules and filtering and didn't test
all the code path.

This patch introduces a new context manager to do request as admin, and
switches to a non-admin user by default. This also fixes a lot of
problem in the code disovered by this change.

Change-Id: I9328f425a8a678685ecc8e28ea6335ffcd3dcbf5