WIP start adding gabbi tests to gnocchi
This doesn't do very much, I'm posting it just to get some first
glance review and sanity checkin on how I'm doing the database
setup.

I've made two new tox targets: py27-gabbi, and py34-gabbi. The
former users a mysql indexer, the latter a postgresql.

The only actual test that happens so far is requesting / which
gets some lovely text.

No exercising of the storage engine at all yet, but the idea is that
the file driver will be used and a temporary directory. That will need
some tweaking as well as perhaps some cleanup.

Change-Id: Ic09743af5a19c04cecf28bbc1a42e841e142b600