WIP start adding gabbi tests to gnocchi
This doesn't cover everything yet, I'm posting it just to get some
first glance review and sanity checking on how I'm doing the database
setup.

The gabbi tests are integrated with the other tests such that
existing tox targets will call them. This requires making
a ConfigFixture that each gabbi yaml file uses. The upshot of this
is that database and conf handling is per file and relatively
clear cut.

It is also possible to run the gabbi tests with their own tox
target: tox -epy27-gabbi.

There are three test files thus far:

* base.yaml requests / and makes sure there is a world
* archive.yaml covers some basic archive policy handling
* resource.yaml starts investigating resource handling.
* metric.yaml starts investigating metric handling.

xfails are used where necessary to ride over bad behavior.

No exercising of the storage engine at all yet, but the idea is that
the file driver will be used and a temporary directory. That will need
some tweaking as well as perhaps some cleanup.

Change-Id: Ic09743af5a19c04cecf28bbc1a42e841e142b600