Use werkzeug to run the developement API server
wsgi.simple_server in a mono threaded process that can handle only
5 requests at a time.

If we use the Gnocchi Ceilometer dispatcher, this one can create
a ton of HTTP connections and most of them received  "connection denied"
errors.

Even the doc recommands to setup Gnocchi through an other WSGI services
like Apache 'mod_wsgi', we can provide a better testing API server.

So this patch changes the default HTTP server to the werkzeug one with
a autodiscovery of number of workers that we can use.

The client queue of werkzeug is 128, so on a 4 cpus machine, gnocchi-api
can now handle 512 connections instead of 5.

Also the change adds references of how to deploy pecan application in
the documentation.

Change-Id: Idbcd8ea3e4d5b67593f9b5a9d69be4316ed39ded