Base infrastructure to support gabbi tests
Provides the code necessary to run gabbi tests hosted in yaml files
in gnocchi/tests/gabbi/gabbits. The provided tests do very little
other than prove that it is working. Subsequent patches may add
more rigorous and vigorous tests by adding yaml files.

The gabbi tests are either run as part of other tox targets the
include a database, or by a 'py27-gabbi' target. If not indexer is
available, those gabbi tests which require the ConfigFixture (all
of them because of the way the api server starts up) will be skipped.

Change-Id: I9fece3b608847177ab30cff082ecee0d33e34cc5