Cover the basic resource API with gabbi tests
These tests cover the ResourceController and ResourcesController
classes. This is not to suggest that they mean those API endpoints
are fully tested but rather than the code is used. Edge cases and
issues with authorization will need to be handled in subsequent
patchsets.

Note that several of the included tests are flagged as being
expected to fail. Where this is the case there are links to
launchpad bugs.

Missing from this is removal of the corresponding pre-existing
WebTest tests.

Change-Id: Idecec1c97da327b599901fd458ea3a593f5967f7