Gnocchi driver for KairosDB with Cassandra
The driver allows to:
* create metrics
* add raw and pre-aggregated measures
* get raw or preaggregated measures
* delete measures
* delete metrics

There is a write delay in KairosDB which default value is
equal to 1000milliseconds. This value has been changed to
20 milliseconds. Driver’s config parameter 'kairosdb_write_delay'
is set to this value in seconds. To change write delay value in
KairosDB, change this line:
'kairosdb.datastore.cassandra.write_delay=20'
in conf/kairosdb.properties file.

Change-Id: Ifa326022dc94b26711bab28d840be034a4dc0a2a