rest: remove PATCH schemas
The REST API currently puts a lot of constraint and difference between
resource creation and update. I don't think it's currently a
requirement, and that complicates a lot of the code of resource
declaration by duplicating their fields.

This patches makes use of new voluptuous feature that allow us to use
the same resource schema for both creation and update of a resource.

Change-Id: I9a5f6f003f13122188324701d297f22628c19fdf