sqlalchemy: enhance QueryTransformer
This enhances QueryTransformer in sqlalchemy based on the recent work
that have been done with the QueryMeasure in _carbonara.

It also maps all the indexer exception under a generic IndexerException
query so the REST API can only catch this one and serialize it down to
the client with a 400 error. We know that these exception are not a
problem with the driver, but with the client sending garbage.

Change-Id: I493e8cfcf503329bee92acead14f9c66b475dbde