indexer: return resource objects rather than dict
We used to return dict from the indexer, and then convert them to
object, and then sometimes encoding them as JSON.

We know make the indexer return directly the object that has the right
interface to be directly encoded as JSON. This reduces the number of
objects we have to create down the path and improves performance.

This is done for all methods that handles resources.

Change-Id: I0e6034083f1fd8f55d924dd0eaaee17ffb35c691