rest: enhance metric retrieval in controllers
This fetches the metric information earlier in the API, so it avoids doing
more than one request on the indexer to retrieve a metric info. The
upside is that we detect also non existent metric earlier.

The downside is that we have to fetch the details of a metric (i.e. the
archive policy detail) on each call because most controllers expect to
have these details available. That shouldn't be a large issue, it ends up
simplifying the API a bit since you don't have to explicitely request
the details of a metric anymore.

Change-Id: If0af07fdb83b05e6172141ebc4687cf4de7e51c9