Add support for ArchivePolicy Rule Endpoints
The goal of this is to expose an end point where
user could define an archive policy rule mapping
to a metric pattern. There could be multiple rules
defined. This rule is used when a metric is created
to automatically assign a policy to the metric based
on matching pattern defined in the rule.

There will be a subsequent patch that leverages this
in the metric api.

Change-Id: I8a71f25fe25c4be20d14d695c52fa7cb6efcb14b