Let WebOb read the body of a PATCH when creating docs
webob had a bug wherein the body was not read on a PATCH. A pull
request was submitted:

    https://github.com/Pylons/webob/pull/184

and a new release with that change was released recently, breaking
fresh 'tox -e docs' jobs.

This changeset just requires the working WebOb. This may present
packaging issues, so tidier solutions welcome.

Change-Id: I650a3ca3ae8381c556457c4504460c2b7f8a8d60