Only read body on PATCH if it is not already read.
webob had a bug wherein the body was not read on a PATCH. A pull
request was submitted:

    https://github.com/Pylons/webob/pull/184

and a new release with that change was released recently, breaking
fresh 'tox -e docs' jobs.

This changeset makes the guard for PATCH in gendoc check for a value
on body before attempting to read. If there's something there we
have the modern webob.

Change-Id: I650a3ca3ae8381c556457c4504460c2b7f8a8d60