Use project instead of user for activity filter
We choice to use an user for filtering out the gnocchi activity on swift
because it looks more logical that the application is represented by a
user from the keystone point of view.

But unfortunatly, some samples generated by gnocchi activity and ingested
by the gnocchi dispatcher can have the user_id field empty like
objectstore samples created by the swift pollster.

So this patch uses a project instead of an user to filter the activity of
gnocchi.

Change-Id: If57f905580dfacb099b39524cdb6175f5c2c581f