rest: fix access to metric for owned resources
If a user owns a resource (= user_id/project_id set to its UUID) and
this resource has metrics linked to it, the user cannot access these
metrics because he's not created those metrics. Let's make sure we rely
on the metric.resource.project_id to check an access too.

Change-Id: Ibd38da9b596e4c74e477f9f7be14fb563907d7b1