Run functional tests using virtualenvs in devstack
The functional tests are failing because keystone scripts traverse
their own requirements when they start up and die if there is no
match. Since we wish to use requirements that are uncapped we can
constrain the impact of other requirements setting USE_VENV=True
when setting up the devstack.

Gnocchi will be run its own virtualenv as well. The call to
install_gnocchi is replaced with stack_install_service gnocchi which
does the necessary environment variable setting to ensure that pip
will use the virtualenv when doing installs.

Change-Id: Ib1e0cd087d11b317193d75110795e70a9f828a60