Remove redundancy in ArchivePoliciesController tests
Except for aggregation related tests, duplicate /v1/archive_policy
related tests from test_rest into gabbi tests to remove redundancy.

The initial plan was to just remove tests from test_rest that had
duplicates under gabbi, but it was easier to reconcile into one
source of authority.

Note the addition of two xfail-ing tests in archive.yaml. These fail
on python3 but not python2. The issue is in pecan as described in
the linked bug report.

Change-Id: I4a9d9f1f2a1ced072f91263bf8e819cc798b123a