[WIP] Cover unicode handling in mod_wsgi setup
On some setups the default encoding that will be present in the
Python running under mod_wsgi will not be UTF-8 compatible and will
cause errors when decoding headers in the WSGI environment.

This commit is testing that it will _fail_ as expected. The fix and
an updated commit will come after confirming failure.

Change-Id: I6f470f8ab311f8debd1758840dd74b056abbc9ea