Make measures processing asynchronously
This patches split the measures processing for the Carbonara in
two-phases: first the measures are written as fast as possible, and then
a function is responsible to process them in batch.

Currently this is still done synchronously for the API, but a next patch
will split that into a secondary daemon.

This is not implemented for Ceph.

Change-Id: I08a877bbcf7abb034188c5710ba79f5f81a5a9a0