Avoid conflicts with ceilometer.conf in dispatcher tests
This makes sure a critical configuration value
(filter_service_activity) used in the ceilometer dispatcher tests
is explicitly set in the setUp of the test. Otherwise a conflicting
value in /etc/ceilometer/ceilometer.conf can cause tests to fail.

Efforts to make a solution which just ignores
/etc/ceilometer/ceilometer.conf were confounded by too much "helpful"
magic in oslo_config. Given that this dispatcher will be leaving the
gnocchi repo soonish, a long trip down a rabbit hole was skipped.

Change-Id: I8f270183d6c38de0b4088166ddc2da687576a752
Closes-Bug: #1463889