Add a thread in ConfigFixture to do metricd
Without this the new async processing of measures will cause tests
that check to see if measures were created will fail.

The thread runs the process_measures method in a tight loop.

Meanwhile gabbi >=0.100.0 provides a 'poll' function so that tests
will try a few times for what they are looking for.

The tests in resource.yaml have been cleaned up to insure that they
are testing a specific resource and its measures not an ambiguous
last-in-the-list member of resources.

Change-Id: Id18f4a607ad910fdf6738b225811073c89a03508