Allow gnocchi API at /metric
One of the devstack goals is to get all the services running at
endpoints that are under the same host and port, for example:

   http://api.example.com/metric
   http://api.example.com/telemetry
   http://api.example.com/identity

This changeset enables that.

GNOCCHI_SERVICE_PREFIX defaults to /metric

The service will by default run on port 80 at /metric/v1/...

If GNOCCHI_SERVICE_PORT is defined (undefined by default) the server
will run without a prefix, on the stated port.

The gabbi-based live test will use the server addressed by
GNOCCHI_SERVICE_URL. If that is not set, the live test won't run.

There's additional work to do here to get things fully aligned with
the goals but this gets things moving in the right direction. One of
the next steps is for the wsgi script to not require explicit
installation from the devstack plugin. Ideally the pip install
process would accomplish this.

Change-Id: Iec2924b5460afe2ebeafab7c82eb4ddf184f2c99