Allow gnocchi API at /gnocchi
One of the devstack goals is to get all the services running at
endpoints that are under the same host and port, for example:

   http://api.example.com/gnocchi
   http://api.example.com/ceilometer
   http://api.example.com/keystone

This changeset enables that.

GNOCCHI_SERVICE_PREFIX defaults to /gnocchi

The service will run at:

* GNOCCHI_SERVICE_PORT (8041) at /v1/...
* port 80 at /gnocchi/v1/...

and the official keystone endpoint will be the port 80 based URL.

The gabbi-based live test will use prefix if it is provided (but will fail
until a bug in gabbi is fixed).

There's additional work to do here to get things fully aligned with
the goals but this gets things moving in the right direction. A next
step is for the wsgi script to not require explicit installation.

Change-Id: Iec2924b5460afe2ebeafab7c82eb4ddf184f2c99