rest: returns 404 when controller does not exists
Since version 0.9.0, pecan returns 405 when no
method or routing object are found into a rest.Controller.

This change removes inheritance of pecan.rest.Controller
when the class is just a routing controller to lets pecan
returning the correct status code.

It also changes the lookup method for the ResourcesController and
SearchResourceController to appropriatly returns 404 when the
resource_type does not exists.

Closes bug: #1463102

Change-Id: I4661baed4fbf58b3c1a00cf6a292b95122321693