Switch to facade
WIP.  the SQLAlchemyIndexer is a class with methods, which is
not the standalone-function based calling style that the new
enginefacade was designed for; additionally, it does not use
the global engine scope and instead has its own, which is
even disconnectable.  So many changes will be needed in
enginefacade to support these patterns.  Many more tests
are passing now however there are still some session
related problems.

Change-Id: I4e60cca3434e4290145a6a38ce6f27a6c7b1d2be
Signed-off-by: Julien Danjou <julien@danjou.info>