storage: remove create_metric()
This changes the storage driver API so that there's no need to create
metric before adding metrics. We let the indexer be responsible for the
status of a metric to be existing or non existing, and we allow the
storage driver to create metric on the fly.

Pros:

1. Speeds up the metric creation process as we do not need to reach the
   storage to create a container + empty archives
2. Avoid us for a COMMIT/ROLLBACK step we would have to handle in the
   metric creation process, in the case where creating a metric either
   storage or indexer would fail – we would have to rollback the
   creation that did not fail.

Cons:

1. We now acknowledge that it's more difficult to have an autonomous
   storage driver working without an indexer. While this was no obvious
   before Gnocchi 1.0.0, it's getting pretty clear now that we don't
   want to bypass the indexer as it's anyway responsible for things such
   as RBAC. And we can anyway cache him.

This will help fixing bug #1483634 since we'll be able to create metric
only in the indexer, doing that in only one single transaction.

This change also handles correctly deletion of unprocessed measures for
metrics that has been deleted.

Change-Id: I81ff5ca5540a8a02d378c289bed611a1329f9325
Related-Bug: #1483634