Rudimentary support for influxdb in devstack plugin
This is bare bones and gets the job done without too much
complexity. It relies, for the time being, on the influxdb
distributors not changing their paths and us keeping those paths up
to date.

The packages provided by influxdb are used because the packags from
the distros are currently too old. Because we are using these weird
packages, the start and stop handling for influxdb is a bit strange.
When the distros are up to date, the _gnocchi_install_influxdb
function should be updated

Note that influxdb is running _without_ access control.

Change-Id: If5b3f05c9e4044efd2a5e4c5b7231d0b5856d9ca