WIP: Introduce a resource id translation shim
Proof of concept potential solution to the problem described below.
This may be completely insane but since it was easy enough to
implement I thought it worthwhile getting it out for demonstration.

In the openstack environment resource ids are a mismash of UUIDs
(yay!) and other things (boo!). When gnocchi receives a resource id
that is not a UUID, it drops it on the floor with a 400 error.

In some sense this is the exact right thing to do. In others it is
entirely impractical given the use context.

This change introduces a shim which accepts any string as a resource
id, either in the URL or in the Resource voluptuous Schema and does
a transformation on it to turn it into a UUID. That transformation
is:

* check if the string <= 255 in lenght (ValueErrr is not)
* do a sha256 hash of the string
* take the first 32 characters of the hexdigest() to create a UUID
* use the resulting UUID as the local id for the resource

The transform is done for both writes and reads. This is necessary
because we have no control over the caller.

This lack of control is the same reason this change is being
done in gnocchi rather than the ceilometer dispatcher. Gnocchi users
will except to be able to access data in the Gnocchi API by the IDs
they are familiar with. We could consider putting the shim on the
client side but that would demand that everyone use the clients we
make which is a pretty heavy demand.

To make this a complete solution additional testing of more methods
and more URIs would be required. Pushing out for review first.

Concerns with this way of doing things:

* It violates Gnocchi's integrity as an independent thing for the
  sake of OpenStack's problem.
* It makes it hard to dig around in the datastore when there are
  problems (mismatch between natural key and database key).
* Doing a lot of sha256 hashing could get tiring after a while.
* No rigorous evaluation of the potential for collision has yet been
  done.

Pros:

* It keeps UUIDs as the primary identifiers internally.
* If people know about some of the cons they might consider fixing
  their ids.
* It's less intrusive than providing a persistent map between
  OpenStack key and Gnocchi key.
* It requires no indexer-database migration (which a change in id
  structure would).

Probably plenty more pros and cons.

Change-Id: Ibfc58d6f239a0474cbb3ccd9a6229393ada77e8f
Related-Bug: #1480346