Retry on initial contact with swift
If gnocchi and keystone start up at the same time the swift backend
may experience authentication errors while keystone is coming up. To
be more resilient in the face of this, the intial put_container call
used at swift+carbonara startup will retry for a minute if it
receives an authorization error.

Due to clumsy exception handling in swiftclient, the retry logic is
less robust than it should be.

Change-Id: I1557588bdb1d65d7c845356f2f63dc71a0aad279
Partial-Bug: #1493060