rest: use a fork friendly app with werkzeug
Unlike many wsgi servers like apache mod_wsgi or uwsgi.

With werkzeug, the gnocchi application is created before
the server fork itself to creates worker.
This have some issues, storage/indexer connection are created
and leaves on the main process for no reason.
And the ceph storage driver just doesn't support to be
forked once the connection to ceph have been established.

This change creates a class that lazy load the application for
werkzeug and fixes these issues.

Closes-bug: #1499115

Change-Id: I398e289767e2e267466bcf882a5bb013620949d0