Fix InfluxDB tests
Now, InfluxDB tests are failing via Timeout error on writing
or getting data. It is caused by internal multithreading locks which
relative to measurements dropping. Also, these locks happen then
we drop measurements due the active writing measures to the current db.

So, in this patchset we avoid to drop measurements and expect that it
will deleting via retention.
For every test we use new database and it allows to use
isolated data for every test.

Also, in this CR timestamp query format have been changed for
compatibility with current Gnocchi tests.

Change-Id: I324a8a1f1b767ecc1a00c3b73a4612a4467842ef
Relative-to: bug #1488027