tests: increase MySQL max connections
Depending on the number of tests running in parallel, the number of
MySQL connection used can go pretty high since it's used both by the
index and by tooz.

Let's bump this limit since on the gate we have 8 CPU and therefore up
to 8 process running tests with each several threads trying to lock
various metric for upgrade or writing.

Change-Id: I7d6d5dba3975d8359bc1b28f4804d642361687de