Checks percent_of_overlap when one boundary is set
Aggregation across metrics have different behavior depending
on if boundary are set and if needed_percent_of_overlap is set.

If boundaries are not set, Carbonara makes the aggregation only with points
at timestamp present in all timeseries.

But when boundaries are set, Carbonara expects that we have certain
percent of timestamps common between timeseries, this percent is controlled
by needed_percent_of_overlap (defaulted with 100%).

This change fixes a weird behavior when only one boundary is set,
the needed_percent_of_overlap wasn't checked.

Change-Id: Ifc85a9004e864d14d42fd482ec144e4d27dd615b
Closes-bug: #1522434