storage: limit the number of unprocessed measures return
Currently, all the unprocessed measures are returned which can build a
very large list if you have millions of unprocessed measures waiting –
ending up in a MemoryError.

To avoid that, let's read those unprocessed measures in smaller batches,
with a more-or-less random size (between 100 and 200) so that all
workers end up desynchronized and only at worst read slightly
overlapping batches.

Closes-Bug: #1533793
Change-Id: I3a0b726aae14a17a23a365babc1a2537fb4d1052