Implements resource attribute string
This change allows to create resource attributes
of type string.

The choosen solution is that the indexer is responsible
to providing voluptuous schemas and build resource columns
according the the voluptuous schema that it proposes.

Another alternative could be that rest API provides
jsonschema for each times and indexer is only
responsible for the schema to sql column convertion.

Blueprint resource-type-rest-api
Change-Id: I7877b6ea97dc70f3629e63abe5ef1ddf61d200b3