swift, ceph: take smaller batch of new measures to process
When checking for metrics that have new measures to be processed, the
current code retrieves all the measures backlog and then transforms it
into a set of all metric needing processing.

While correct and fancy, this can lead to situation where the Ceph
drivers lists millions of new measures objects to process, and construct
a list of this size with a set of a few thousands metric. This is slow,
takes a lot of memory, and turns out to be not strictly necessary.

This patches limit the size to the default listing limit in Swift and to
128 metrics in Ceph. The file driver is unaffected, as os.listdir()
which is being used cannot benefit from any limit: it returns a list,
not an iterator.

Change-Id: If2a546b0c9abeb9fdc0abaea4e22272b5d5241b1
Closes-Bug: #1536909