carbonara: serialize AggregatedTimeSerie using RLE, XOR encoding and LZ4
This changes the serialization format of AggregatedTimeSerie to use a
double-delta RLE encoding for the timestamps followed by a LZ4
compression. The values are encoded using an XOR based encoding.

This reduces the storage to a worst case scenario of ~7.3 bytes per
datapoint, and can compress data down to 0.2 byte per value for simple
discrete values.

Change-Id: Ie9cc8bf036569db5bfc3213bc16145bc16c64108