carbonara: serialize AggregatedTimeSerie using RLE and LZ4
This changes the serialization format of AggregatedTimeSerie to use a
double-delta RLE encoding for the timestamps and a LZ4 compression for
the whole archive.

This reduces the storage to a worst case scenario of 9 bytes per
datapoint, and can compress data down to 0.02 byte per value for simple
discrecte values.

Change-Id: Ie9cc8bf036569db5bfc3213bc16145bc16c64108